Teams: add xiongchiamiov
For better or for worse, I'm on the team, so we've got to publicize it.

This is basically just a copy of #47, but with less spiders.  Or spider-fear, I
guess.

Thanks to Wolfram Alpha for the inches -> meters conversion, and telling me
that it's about 7.8 time the inner radius of an NBA basketball rim.

I seem to be pretty average here:

    [$]> grep top reddit_about/data/team.json | awk '{print $2}' | sort -n | uniq -c